Create billing.php
Modified version of the original PMPro billing page preheader.  Had to modify this to work with Authorize.net CIM, multiple membership levels (instead of one)... and so it's not updating Authorize.net ARB subscriptions (because the subscriptions are stored in a custom database table so we can do our own billing against a CIM payment profile instead - because ARB and CIM are still not compatible with each other yet).